+++ src/main/java/com/travel/role/domain/room/entity/EssentialCategory.java
@@ -1,4 +1,4 @@
-package com.travel.role.domain.room.domain;
+package com.travel.role.domain.room.entity;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
