+++ src/main/java/com/travel/role/domain/room/entity/RoomRole.java
@@ -1,4 +1,4 @@
-package com.travel.role.domain.room.domain;
+package com.travel.role.domain.room.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
