+++ src/main/java/com/travel/role/domain/room/repository/ParticipantRoleRepository.java
@@ -1,9 +1,9 @@
-package com.travel.role.domain.room.dao;
+package com.travel.role.domain.room.repository;
 
-import com.travel.role.domain.room.domain.ParticipantRole;
-import com.travel.role.domain.room.domain.Room;
-import com.travel.role.domain.room.domain.RoomRole;
-import com.travel.role.domain.user.domain.User;
+import com.travel.role.domain.room.entity.ParticipantRole;
+import com.travel.role.domain.room.entity.Room;
+import com.travel.role.domain.room.entity.RoomRole;
+import com.travel.role.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
