+++ src/main/java/com/travel/role/domain/room/repository/RoomRepository.java
@@ -1,10 +1,10 @@
-package com.travel.role.domain.room.dao;
+package com.travel.role.domain.room.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.travel.role.domain.room.domain.Room;
+import com.travel.role.domain.room.entity.Room;
 
 public interface RoomRepository extends JpaRepository<Room, Long>, RoomQuerydsl {
     boolean existsByRoomInviteCode(String inviteCode);
