+++ src/main/java/com/travel/role/domain/room/repository/TravelEssentialQuerydsl.java
@@ -1,9 +1,9 @@
-package com.travel.role.domain.room.dao;
+package com.travel.role.domain.room.repository;
 
 import java.util.List;
 import java.util.Map;
 
-import com.travel.role.domain.room.domain.EssentialCategory;
+import com.travel.role.domain.room.entity.EssentialCategory;
 import com.travel.role.domain.room.dto.TravelEssentialResDTO;
 
 public interface TravelEssentialQuerydsl {
