+++ src/main/java/com/travel/role/domain/room/service/TravelEssentialService.java
@@ -6,18 +6,18 @@ import java.util.Map;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.travel.role.domain.room.dao.RoomParticipantRepository;
-import com.travel.role.domain.room.dao.RoomRepository;
-import com.travel.role.domain.room.dao.TravelEssentialRepository;
-import com.travel.role.domain.room.domain.EssentialCategory;
-import com.travel.role.domain.room.domain.Room;
-import com.travel.role.domain.room.domain.TravelEssentials;
+import com.travel.role.domain.room.repository.RoomParticipantRepository;
+import com.travel.role.domain.room.repository.RoomRepository;
+import com.travel.role.domain.room.repository.TravelEssentialRepository;
+import com.travel.role.domain.room.entity.EssentialCategory;
+import com.travel.role.domain.room.entity.Room;
+import com.travel.role.domain.room.entity.TravelEssentials;
 import com.travel.role.domain.room.dto.TravelEssentialCheckReqDTO;
 import com.travel.role.domain.room.dto.TravelEssentialDeleteReqDTO;
 import com.travel.role.domain.room.dto.TravelEssentialReqDTO;
 import com.travel.role.domain.room.dto.TravelEssentialResDTO;
-import com.travel.role.domain.user.dao.UserRepository;
-import com.travel.role.domain.user.domain.User;
+import com.travel.role.domain.user.repository.UserRepository;
+import com.travel.role.domain.user.entity.User;
 import com.travel.role.domain.user.exception.RoomInfoNotFoundException;
 import com.travel.role.domain.user.exception.UserInfoNotFoundException;
 import com.travel.role.domain.user.exception.UserNotParticipateRoomException;
