+++ src/main/java/com/travel/role/domain/room/service/WantPlaceService.java
@@ -1,17 +1,17 @@
 package com.travel.role.domain.room.service;
 
-import com.travel.role.domain.room.dao.ParticipantRoleRepository;
-import com.travel.role.domain.room.dao.RoomParticipantRepository;
-import com.travel.role.domain.room.dao.RoomRepository;
-import com.travel.role.domain.room.dao.WantPlaceRepository;
-import com.travel.role.domain.room.domain.Room;
-import com.travel.role.domain.room.domain.RoomRole;
-import com.travel.role.domain.room.domain.WantPlace;
+import com.travel.role.domain.room.repository.ParticipantRoleRepository;
+import com.travel.role.domain.room.repository.RoomParticipantRepository;
+import com.travel.role.domain.room.repository.RoomRepository;
+import com.travel.role.domain.room.repository.WantPlaceRepository;
+import com.travel.role.domain.room.entity.Room;
+import com.travel.role.domain.room.entity.RoomRole;
+import com.travel.role.domain.room.entity.WantPlace;
 import com.travel.role.domain.room.dto.WantPlaceDTO;
 import com.travel.role.domain.room.dto.WantPlaceRequestDTO;
 import com.travel.role.domain.room.dto.WantPlaceResponseDTO;
-import com.travel.role.domain.user.dao.UserRepository;
-import com.travel.role.domain.user.domain.User;
+import com.travel.role.domain.user.repository.UserRepository;
+import com.travel.role.domain.user.entity.User;
 import com.travel.role.domain.user.exception.PlaceInfoNotFoundException;
 import com.travel.role.domain.user.exception.RoomInfoNotFoundException;
 import com.travel.role.domain.user.exception.UserInfoNotFoundException;
