--- src/main/java/com/travel/role/domain/user/domain/Provider.java
@@ -1,6 +1,0 @@
-package com.travel.role.domain.user.domain;
-
-public enum Provider {
-	local,
-	google,
-}
