+++ src/main/java/com/travel/role/domain/user/entity/Provider.java
@@ -1,0 +1,6 @@
+package com.travel.role.domain.user.entity;
+
+public enum Provider {
+	local,
+	google,
+}
