+++ src/main/java/com/travel/role/domain/user/entity/Role.java
@@ -1,4 +1,4 @@
-package com.travel.role.domain.user.domain;
+package com.travel.role.domain.user.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
