+++ src/main/java/com/travel/role/domain/user/entity/User.java
@@ -1,4 +1,4 @@
-package com.travel.role.domain.user.domain;
+package com.travel.role.domain.user.entity;
 
 import java.time.LocalDate;
