+++ src/main/java/com/travel/role/global/auth/service/handler/OAuth2SuccessHandler.java
@@ -16,8 +16,8 @@ import org.springframework.security.web.authentication.AuthenticationSuccessHand
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.travel.role.domain.user.dao.UserRepository;
-import com.travel.role.domain.user.domain.User;
+import com.travel.role.domain.user.repository.UserRepository;
+import com.travel.role.domain.user.entity.User;
 import com.travel.role.global.auth.dto.TokenMapping;
 import com.travel.role.global.auth.service.RefreshTokenCookieProvider;
 import com.travel.role.global.auth.service.TokenProvider;
