+++ src/test/java/com/travel/role/domain/room/service/RoomServiceTest.java
@@ -15,18 +15,18 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import com.travel.role.domain.room.dao.ParticipantRoleRepository;
-import com.travel.role.domain.room.dao.RoomRepository;
-import com.travel.role.domain.room.domain.Room;
+import com.travel.role.domain.room.repository.ParticipantRoleRepository;
+import com.travel.role.domain.room.repository.RoomRepository;
+import com.travel.role.domain.room.entity.Room;
 import com.travel.role.domain.room.dto.MakeRoomRequestDTO;
 import com.travel.role.domain.room.exception.AlreadyExistInRoomException;
 import com.travel.role.domain.room.exception.InvalidInviteCode;
 import com.travel.role.domain.room.exception.InvalidLocalDateException;
 import com.travel.role.domain.room.exception.UserHaveNotPrivilegeException;
-import com.travel.role.domain.user.dao.UserRepository;
-import com.travel.role.domain.user.domain.Provider;
-import com.travel.role.domain.user.domain.Role;
-import com.travel.role.domain.user.domain.User;
+import com.travel.role.domain.user.repository.UserRepository;
+import com.travel.role.domain.user.entity.Provider;
+import com.travel.role.domain.user.entity.Role;
+import com.travel.role.domain.user.entity.User;
 import com.travel.role.domain.user.exception.UserInfoNotFoundException;
 import com.travel.role.global.auth.token.UserPrincipal;
 import com.travel.role.global.exception.ExceptionMessage;
